src/requirements.txt
fastapi
uvicorn
passlib
python-multipart==0.0.5
httpx
python-jose

# dev
pytest
requests
